+++ plat/xilinx/common/pm_service/pm_ipi.c
@@ -299,12 +299,12 @@ void pm_ipi_irq_clear(const struct pm_proc *proc)
 
 uint32_t pm_ipi_irq_status(const struct pm_proc *proc)
 {
-	int32_t ret;
+	uint32_t ret;
 	uint32_t result = (uint32_t)PM_RET_SUCCESS;
 
 	ret = ipi_mb_enquire_status(proc->ipi->local_ipi_id,
 				    proc->ipi->remote_ipi_id);
-	if (((uint32_t)ret & IPI_MB_STATUS_RECV_PENDING) != 0U) {
+	if ((ret & IPI_MB_STATUS_RECV_PENDING) != 0U) {
 		result = IPI_MB_STATUS_RECV_PENDING;
 	}
